Add tolerance to minimiziation
Problem 'PORTFL2' is a 12 dimensional problem with an equality constraint that forces the sum of the elements of x to be 1.

The provided x0 consists of 12 values of 1/12, but in single precision. When the minimization routine runs, it exits without doing anything because the sum is pretty close to 1.

However for algorithms like LINCOA the starting point is still infeasible and this can cause problems. In this case the starting point will always be infeasible since 1/12 cannot be represented in finite precision floating point, but by adding the tolerance we can reduce the constraint violation.